well it seems to work, seg faults randomly

+++ berkeleyDB/simpledb.h
@@ -33,7 +33,7 @@ struct character{
 struct db_log_record{
     uint64_t time;
     uint32_t XID;
-    uint8_t type;
+    uint32_t type;
     uint32_t key;
     //uint32_t offset;
     uint32_t data_length;
